VGG16 model trained
This commit is for task 1 part ii, training a cnn using a pretrained model of VGG16.
This model uses 200 images in training, 40 in test. Also used Adam learning rate or 0.0001.
Matthias is working on optimization of this at the moment so we may update this class but it works well enough for now.
README also added for Task 1.